+++ src/design/plone/contenttypes/upgrades/upgrades.py
@@ -870,3 +870,20 @@ def to_5210(context):
     if "volto.preview_image" not in behaviors:
         behaviors.append("volto.preview_image")
         fti.behaviors = tuple(behaviors)
+
+
+def to_5220(context):
+    """
+    Reindex Venues
+    """
+    logger.info("Reindex SearchableText for Venue items.")
+    catalog = api.portal.get_tool("portal_catalog")
+    i = 0
+    brains = catalog(portal_type="Venue")
+    tot = len(brains)
+    for brain in brains:
+        i += 1
+        if i % 500 == 0:
+            logger.info("Progress: {}/{}".format(i, tot))
+        obj = brain.getObject()
+        obj.reindexObject(idxs=["SearchableText", "object_provides"])